slight changes
added nav bar to each page and made a bunch of cosmetic changes. added the button to click to sign up for the newsletter and also to view the link tree. also added the icons to see the social media. changed the color to match the color palette carla uses on the instagram.